Move the menu packages.
