+++ src/controller/SubViewLoader.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 10/12/2019
+ * Copyright (c) 11/12/2019
  *
  * Auteurs :
  *      - Behm Guillaume
@@ -10,8 +10,8 @@
 
 package controller;
 
-import controller.menus.mainMenu.MainMenuController;
-import controller.menus.mainMenu.settings.SettingsController;
+import controller.menus.MainMenuController;
+import controller.menus.settings.SettingsController;
 import javafx.stage.Stage;
 
 /**
