+++ src/controller/ViewController.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 10/12/2019
+ * Copyright (c) 11/12/2019
  *
  * Auteurs :
  *      - Behm Guillaume
@@ -10,7 +10,7 @@
 
 package controller;
 
-import controller.menus.mainMenu.MainMenuController;
+import controller.menus.MainMenuController;
 import javafx.stage.Stage;
 
 import java.util.ResourceBundle;
