+++ src/controller/menus/MainMenu.fxml
@@ -14,7 +14,7 @@
 <?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.layout.VBox?>
 <VBox alignment="CENTER" prefHeight="400.0" prefWidth="252.0" styleClass="vbox" xmlns="http://javafx.com/javafx/9.0.1"
-      xmlns:fx="http://javafx.com/fxml/1" fx:controller="controller.menus.mainMenu.MainMenuController">
+      xmlns:fx="http://javafx.com/fxml/1" fx:controller="controller.menus.MainMenuController">
     <children>
         <AnchorPane prefHeight="200.0" prefWidth="200.0"/>
         <VBox alignment="CENTER" prefHeight="200.0" prefWidth="100.0" spacing="10.0">
